Move ping command to essentials, info -> playerinfo
