+++ core/src/main/java/com/eternalcode/core/EternalCore.java
@@ -32,6 +32,7 @@ import com.eternalcode.core.database.DatabaseManager;
 import com.eternalcode.core.database.NoneRepository;
 import com.eternalcode.core.database.wrapper.HomeRepositoryOrmLite;
 import com.eternalcode.core.database.wrapper.IgnoreRepositoryOrmLite;
+import com.eternalcode.core.feature.PingCommand;
 import com.eternalcode.core.feature.adminchat.AdminChatCommand;
 import com.eternalcode.core.feature.afk.AfkCommand;
 import com.eternalcode.core.feature.afk.AfkController;
@@ -63,6 +64,9 @@ import com.eternalcode.core.feature.essentials.item.ItemLoreCommand;
 import com.eternalcode.core.feature.essentials.item.ItemNameCommand;
 import com.eternalcode.core.feature.essentials.item.RepairCommand;
 import com.eternalcode.core.feature.essentials.item.SkullCommand;
+import com.eternalcode.core.feature.essentials.playerinfo.OnlinePlayerCountCommand;
+import com.eternalcode.core.feature.essentials.playerinfo.OnlinePlayersListCommand;
+import com.eternalcode.core.feature.essentials.playerinfo.WhoIsCommand;
 import com.eternalcode.core.feature.essentials.speed.SpeedCommand;
 import com.eternalcode.core.feature.essentials.time.DayCommand;
 import com.eternalcode.core.feature.essentials.time.NightCommand;
@@ -79,10 +83,6 @@ import com.eternalcode.core.feature.home.command.SetHomeCommand;
 import com.eternalcode.core.feature.ignore.IgnoreCommand;
 import com.eternalcode.core.feature.ignore.IgnoreRepository;
 import com.eternalcode.core.feature.ignore.UnIgnoreCommand;
-import com.eternalcode.core.feature.info.OnlinePlayerCountCommand;
-import com.eternalcode.core.feature.info.OnlinePlayersListCommand;
-import com.eternalcode.core.feature.info.PingCommand;
-import com.eternalcode.core.feature.info.WhoIsCommand;
 import com.eternalcode.core.feature.privatechat.PrivateChatCommand;
 import com.eternalcode.core.feature.privatechat.PrivateChatPresenter;
 import com.eternalcode.core.feature.privatechat.PrivateChatReplyCommand;
